# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive REST API providing detailed country information and passport visa requirements worldwide. This modern, high-performance service offers extensive data about countries, including demographics, geography, international codes, and visa regulations for various passports.",
        "title": "Atlas - Geographic and Passport Data API by DoROAD",
        "termsOfService": "http://atlas.doroad.io/terms/",
        "contact": {
            "name": "Atlas API Support",
            "url": "https://github.com/DoROAD-AI/atlas/issues",
            "email": "support@doroad.ai"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/DoROAD-AI/atlas/blob/main/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/v2",
    "paths": {
...
        "/passports/visa": {
            "get": {
                "description": "Get visa requirements for a passport holder from one country traveling to another.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passports"
                ],
                "summary": "Get visa requirements between two countries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Origin country code (e.g., USA)",
                        "name": "fromCountry",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination country code (e.g., DEU)",
                        "name": "toCountry",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2.VisaRequirement"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v2.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passports/{passportCode}": {
            "get": {
                "description": "Get visa requirement data for a specific passport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passports"
                ],
                "summary": "Get passport data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Passport code (e.g., USA)",
                        "name": "passportCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v2.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passports/{passportCode}/visas": {
            "get": {
                "description": "Get visa requirements for all destinations for a specific passport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passports"
                ],
                "summary": "Get visa requirements for a passport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Passport code (e.g., USA)",
                        "name": "passportCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v2.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v2.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "v2.VisaRequirement": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string",
                    "example": "USA"
                },
                "requirement": {
                    "type": "string",
                    "example": "90"
                },
                "to": {
                    "type": "string",
                    "example": "DEU"
                }
            }
        }
    }
}